Refactor ModularChangeFamily.invert into a static function
