+++ packages/dds/tree/src/feature-libraries/modular-schema/modularChangeUtils.ts
@@ -1,0 +1,327 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { assert, fail } from "@fluidframework/core-utils/internal";
+import {
+	newChangeAtomIdRangeMap,
+	type ChangesetLocalId,
+	type FieldKey,
+	type FieldKindIdentifier,
+	type RevisionInfo,
+	type RevisionTag,
+	type TaggedChange,
+	type TreeChunk,
+} from "../../core/index.js";
+import { brand, type Mutable, type RangeQueryResult } from "../../util/index.js";
+import { newChangeAtomIdBTree, type ChangeAtomIdBTree } from "../changeAtomIdBTree.js";
+import {
+	CrossFieldTarget,
+	getFirstFromCrossFieldMap,
+	setInCrossFieldMap,
+	type CrossFieldManager,
+	type CrossFieldMap,
+} from "./crossFieldQueries.js";
+import type { FieldChangeHandler } from "./fieldChangeHandler.js";
+import { NodeAttachState } from "./fieldChangeHandler.js";
+import type { FlexFieldKind } from "./fieldKind.js";
+import { genericFieldKind } from "./genericFieldKind.js";
+import {
+	newCrossFieldKeyTable,
+	type CrossFieldKeyTable,
+	type FieldChange,
+	type FieldChangeMap,
+	type FieldId,
+	type ModularChangeset,
+	type NoChangeConstraint,
+	type NodeChangeset,
+	type NodeId,
+} from "./modularChangeTypes.js";
+
+export function hasConflicts(change: ModularChangeset): boolean {
+	return (change.constraintViolationCount ?? 0) > 0;
+}
+
+export function makeModularChangeset(props?: {
+	fieldChanges?: FieldChangeMap;
+	nodeChanges?: ChangeAtomIdBTree<NodeChangeset>;
+	nodeToParent?: ChangeAtomIdBTree<FieldId>;
+	nodeAliases?: ChangeAtomIdBTree<NodeId>;
+	crossFieldKeys?: CrossFieldKeyTable;
+	maxId: number;
+	revisions?: readonly RevisionInfo[];
+	constraintViolationCount?: number;
+	noChangeConstraint?: NoChangeConstraint;
+	noChangeConstraintOnRevert?: NoChangeConstraint;
+	builds?: ChangeAtomIdBTree<TreeChunk>;
+	destroys?: ChangeAtomIdBTree<number>;
+	refreshers?: ChangeAtomIdBTree<TreeChunk>;
+}): ModularChangeset {
+	const p = props ?? { maxId: -1 };
+	const changeset: Mutable<ModularChangeset> = {
+		fieldChanges: p.fieldChanges ?? new Map<FieldKey, FieldChange>(),
+		nodeChanges: p.nodeChanges ?? newChangeAtomIdBTree(),
+		nodeToParent: p.nodeToParent ?? newChangeAtomIdBTree(),
+		nodeAliases: p.nodeAliases ?? newChangeAtomIdBTree(),
+		crossFieldKeys: p.crossFieldKeys ?? newCrossFieldKeyTable(),
+	};
+
+	if (p.revisions !== undefined && p.revisions.length > 0) {
+		changeset.revisions = p.revisions;
+	}
+	if (p.maxId >= 0) {
+		changeset.maxId = brand(p.maxId);
+	}
+	if (p.constraintViolationCount !== undefined && p.constraintViolationCount > 0) {
+		changeset.constraintViolationCount = p.constraintViolationCount;
+	}
+	if (p.noChangeConstraint !== undefined) {
+		changeset.noChangeConstraint = p.noChangeConstraint;
+	}
+	if (p.noChangeConstraintOnRevert !== undefined) {
+		changeset.noChangeConstraintOnRevert = p.noChangeConstraintOnRevert;
+	}
+	if (p.builds !== undefined && p.builds.size > 0) {
+		changeset.builds = p.builds;
+	}
+	if (p.destroys !== undefined && p.destroys.size > 0) {
+		changeset.destroys = p.destroys;
+	}
+	if (p.refreshers !== undefined && p.refreshers.size > 0) {
+		changeset.refreshers = p.refreshers;
+	}
+	return changeset;
+}
+
+// TODO: TFieldData could instead just be a numeric ID generated by the CrossFieldTable
+// The CrossFieldTable could have a generic field ID to context table
+export interface CrossFieldTable<TFieldData> {
+	srcTable: CrossFieldMap<unknown>;
+	dstTable: CrossFieldMap<unknown>;
+	srcDependents: CrossFieldMap<TFieldData>;
+	dstDependents: CrossFieldMap<TFieldData>;
+	invalidatedFields: Set<TFieldData>;
+}
+
+export abstract class CrossFieldManagerI<T> implements CrossFieldManager {
+	public constructor(
+		protected readonly crossFieldTable: CrossFieldTable<T>,
+		private readonly currentFieldKey: T,
+		protected readonly allowInval = true,
+	) {}
+
+	public set(
+		target: CrossFieldTarget,
+		revision: RevisionTag | undefined,
+		id: ChangesetLocalId,
+		count: number,
+		newValue: unknown,
+		invalidateDependents: boolean,
+	): void {
+		if (invalidateDependents && this.allowInval) {
+			const lastChangedId = (id as number) + count - 1;
+			let firstId = id;
+			while (firstId <= lastChangedId) {
+				const dependentEntry = getFirstFromCrossFieldMap(
+					this.getDependents(target),
+					revision,
+					firstId,
+					lastChangedId - firstId + 1,
+				);
+				if (dependentEntry.value !== undefined) {
+					this.crossFieldTable.invalidatedFields.add(dependentEntry.value);
+				}
+
+				firstId = brand(firstId + dependentEntry.length);
+			}
+		}
+		setInCrossFieldMap(this.getMap(target), revision, id, count, newValue);
+	}
+
+	public get(
+		target: CrossFieldTarget,
+		revision: RevisionTag | undefined,
+		id: ChangesetLocalId,
+		count: number,
+		addDependency: boolean,
+	): RangeQueryResult<unknown> {
+		if (addDependency) {
+			// We assume that if there is already an entry for this ID it is because
+			// a field handler has called compose on the same node multiple times.
+			// In this case we only want to update the latest version, so we overwrite the dependency.
+			setInCrossFieldMap(
+				this.getDependents(target),
+				revision,
+				id,
+				count,
+				this.currentFieldKey,
+			);
+		}
+		return getFirstFromCrossFieldMap(this.getMap(target), revision, id, count);
+	}
+
+	public abstract onMoveIn(id: NodeId): void;
+
+	public abstract moveKey(
+		target: CrossFieldTarget,
+		revision: RevisionTag | undefined,
+		id: ChangesetLocalId,
+		count: number,
+	): void;
+
+	private getMap(target: CrossFieldTarget): CrossFieldMap<unknown> {
+		return target === CrossFieldTarget.Source
+			? this.crossFieldTable.srcTable
+			: this.crossFieldTable.dstTable;
+	}
+
+	private getDependents(target: CrossFieldTarget): CrossFieldMap<T> {
+		return target === CrossFieldTarget.Source
+			? this.crossFieldTable.srcDependents
+			: this.crossFieldTable.dstDependents;
+	}
+}
+
+export function newCrossFieldTable<T>(): CrossFieldTable<T> {
+	return {
+		srcTable: newChangeAtomIdRangeMap(),
+		dstTable: newChangeAtomIdRangeMap(),
+		srcDependents: newChangeAtomIdRangeMap(),
+		dstDependents: newChangeAtomIdRangeMap(),
+		invalidatedFields: new Set(),
+	};
+}
+
+export function getRevInfoFromTaggedChanges(changes: TaggedChange<ModularChangeset>[]): {
+	revInfos: RevisionInfo[];
+	maxId: ChangesetLocalId;
+} {
+	let maxId = -1;
+	const revInfos: RevisionInfo[] = [];
+	const revisions = new Set<RevisionTag>();
+	for (const taggedChange of changes) {
+		const change = taggedChange.change;
+		maxId = Math.max(change.maxId ?? -1, maxId);
+		const infosToAdd = revisionInfoFromTaggedChange(taggedChange);
+		for (const info of infosToAdd) {
+			if (!revisions.has(info.revision)) {
+				revisions.add(info.revision);
+				revInfos.push(info);
+			}
+		}
+	}
+
+	const rolledBackRevisions: RevisionTag[] = [];
+	for (const info of revInfos) {
+		if (info.rollbackOf !== undefined) {
+			rolledBackRevisions.push(info.rollbackOf);
+		}
+	}
+
+	rolledBackRevisions.reverse();
+	for (const revision of rolledBackRevisions) {
+		if (!revisions.has(revision)) {
+			revInfos.push({ revision });
+		}
+	}
+
+	return { maxId: brand(maxId), revInfos };
+}
+
+export function revisionInfoFromTaggedChange(
+	taggedChange: TaggedChange<ModularChangeset>,
+): RevisionInfo[] {
+	const revInfos: RevisionInfo[] = [];
+	if (taggedChange.change.revisions !== undefined) {
+		revInfos.push(...taggedChange.change.revisions);
+	} else if (taggedChange.revision !== undefined) {
+		const info: Mutable<RevisionInfo> = { revision: taggedChange.revision };
+		if (taggedChange.rollbackOf !== undefined) {
+			info.rollbackOf = taggedChange.rollbackOf;
+		}
+		revInfos.push(info);
+	}
+	return revInfos;
+}
+
+export function getChangeHandler(
+	fieldKinds: ReadonlyMap<FieldKindIdentifier, FlexFieldKind>,
+	kind: FieldKindIdentifier,
+): FieldChangeHandler<unknown> {
+	return getFieldKind(fieldKinds, kind).changeHandler;
+}
+
+function getFieldKind(
+	fieldKinds: ReadonlyMap<FieldKindIdentifier, FlexFieldKind>,
+	kind: FieldKindIdentifier,
+): FlexFieldKind {
+	if (kind === genericFieldKind.identifier) {
+		return genericFieldKind;
+	}
+	const fieldKind = fieldKinds.get(kind);
+	assert(fieldKind !== undefined, 0x3ad /* Unknown field kind */);
+	return fieldKind;
+}
+
+export interface ConstraintState {
+	violationCount: number;
+}
+
+export function newConstraintState(violationCount: number): ConstraintState {
+	return {
+		violationCount,
+	};
+}
+
+export function updateConstraintsForFields(
+	fields: FieldChangeMap,
+	parentInputAttachState: NodeAttachState,
+	constraintState: ConstraintState,
+	nodes: ChangeAtomIdBTree<NodeChangeset>,
+	fieldKinds: ReadonlyMap<FieldKindIdentifier, FlexFieldKind>,
+): void {
+	for (const field of fields.values()) {
+		const handler = getChangeHandler(fieldKinds, field.fieldKind);
+		for (const [nodeId, inputIndex, _outputIndex] of handler.getNestedChanges(field.change)) {
+			const isInputDetached = inputIndex === undefined;
+			const inputAttachState =
+				parentInputAttachState === NodeAttachState.Detached || isInputDetached
+					? NodeAttachState.Detached
+					: NodeAttachState.Attached;
+
+			updateConstraintsForNode(nodeId, inputAttachState, nodes, constraintState, fieldKinds);
+		}
+	}
+}
+
+function updateConstraintsForNode(
+	nodeId: NodeId,
+	inputAttachState: NodeAttachState,
+	nodes: ChangeAtomIdBTree<NodeChangeset>,
+	constraintState: ConstraintState,
+	fieldKinds: ReadonlyMap<FieldKindIdentifier, FlexFieldKind>,
+): void {
+	const node =
+		nodes.get([nodeId.revision, nodeId.localId]) ?? fail(0xb24 /* Unknown node ID */);
+	if (node.nodeExistsConstraint !== undefined) {
+		const isNowViolated = inputAttachState === NodeAttachState.Detached;
+		if (node.nodeExistsConstraint.violated !== isNowViolated) {
+			node.nodeExistsConstraint = {
+				...node.nodeExistsConstraint,
+				violated: isNowViolated,
+			};
+			constraintState.violationCount += isNowViolated ? 1 : -1;
+		}
+	}
+
+	if (node.fieldChanges !== undefined) {
+		updateConstraintsForFields(
+			node.fieldChanges,
+			inputAttachState,
+			constraintState,
+			nodes,
+			fieldKinds,
+		);
+	}
+}
