+++ packages/dds/tree/src/test/feature-libraries/modular-schema/modularChangesetUtil.ts
@@ -24,11 +24,12 @@ import {
 	type NodeId,
 } from "../../../feature-libraries/index.js";
 import {
-	getChangeHandler,
 	getParentFieldId,
 	normalizeFieldId,
 	// eslint-disable-next-line import-x/no-internal-modules
 } from "../../../feature-libraries/modular-schema/modularChangeFamily.js";
+// eslint-disable-next-line import-x/no-internal-modules
+import { getChangeHandler } from "../../../feature-libraries/modular-schema/modularChangeUtils.js";
 import {
 	newCrossFieldKeyTable,
 	type CrossFieldKeyTable,
